+++ src/Services/HoneypotService.php
@@ -251,10 +251,12 @@ class HoneypotService extends BaseService
             . 'value="' . ($this->isEnhanced() ? $hash : '') . '" '
             . 'name="' . $honeypotName . '" '
             . 'id="' . $honeypotName . '" '
+            . 'aria-hidden="true" '
+            . 'tabindex="-1" '
             . '/>';
 
-        $output = '<div style="position: absolute !important; width: 0 !important; height: 0 !important; overflow: hidden !important;" aria-hidden="true">'
-            . '<label aria-hidden="true" for="' . $honeypotName . '">Leave this field blank</label>'
+        $output = '<div style="position: absolute !important; width: 0 !important; height: 0 !important; overflow: hidden !important;" aria-hidden="true" tabindex="-1">'
+            . '<label aria-hidden="true" tabindex="-1" for="' . $honeypotName . '">Leave this field blank</label>'
             . $output
             . '</div>';